+++ src/containerapp/azext_containerapp/_client_factory.py
@@ -35,7 +35,7 @@ def handle_raw_exception(e):
     stringErr = str(e)
 
     if "WorkloadProfileNameRequired" in stringErr:
-        raise CLIInternalError("Workload profile name is required. use --workload-profile-name.")
+        raise CLIInternalError("Workload profile name is required. Please provide --workload-profile-name.")
 
     if "Unknown properties Name in Microsoft.ContainerApps.WebApi.Views.Version20221101Preview.WorkloadProfile are not supported" in stringErr:
         raise CLIInternalError("Bad Request: Workload profile name is not yet supported in this region.")